+++ packages/umi/index.d.ts
@@ -1,7 +1,21 @@
-export {default as Link} from './link';
-export {default as NavLink} from './navlink';
-export {default as Redirect} from './redirect';
+import { withRouter, RouteComponentProps, RouteProps, match } from 'react-router-dom';
 
-export {default as dynamic} from './dynamic';
-export {default as router} from './router';
-export {default as withRouter} from './withRouter';
+export { default as Link } from './link';
+export { default as NavLink } from './navlink';
+export { default as Redirect } from './redirect';
+
+export { default as dynamic } from './dynamic';
+export { default as router } from './router';
+export { default as withRouter } from './withRouter';
+
+type IncludeRoute = 'component' | 'exact' | 'path';
+
+interface RouteType extends Pick<RouteProps, IncludeRoute> {
+  _title?: string;
+  _title_default?: string;
+}
+
+export interface RouterTypes extends RouteComponentProps {
+  computedMatch?: match;
+  route?: RouteType;
+}
